Fixxed bug where store items were not parsing float values on the item properties

+++ Assets/SocialPlay/Scripts/Core/SPClasses.cs
@@ -145,7 +145,7 @@ public class StoreItem
 public class StoreItemDetail
 {
 	public string propertyName;
-	public int propertyValue;
+	public float propertyValue;
 	public bool invertEnergy;
 }
